# docs/chatgpt_bridge/custom_gpt_openapi.yaml
openapi: 3.1.0
info:
  title: ECHO Local Relay
  version: 1.0.0
  description: Localhost-only ECHO relay action schema for read-only TITAN evidence.
servers:
  - url: https://YOUR-CLOUDFLARE-TUNNEL-HOSTNAME
    description: Replace only after the localhost relay has passed local checks.
components:
  securitySchemes:
    EchoRelayKey:
      type: apiKey
      in: header
      name: X-ECHO-RELAY-KEY
  schemas:
    GenericRelayResponse:
      type: object
      additionalProperties: true
security:
  - EchoRelayKey: []
paths:
...
  /relay/jarvis/ask:
    post:
      operationId: askJarvis
      summary: Ask Jarvis for an evidence-grounded answer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                question:
                  type: string
                context:
                  type: string
              required:
                - question
      responses:
        "200":
          description: Jarvis answer payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericRelayResponse"
  /relay/jarvis/ask/compact:
    post:
      operationId: askJarvisCompact
      summary: Ask Jarvis for a compact evidence-grounded answer.
      description: Compact read-only Jarvis endpoint for Custom GPT use; excludes large details and raw runtime payloads.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                question:
                  type: string
                context:
                  type: string
              required:
                - question
      responses:
        "200":
          description: Compact Jarvis answer payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericRelayResponse"
...